pyproject.toml
[tool.black]
line-length = 79
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
(
  /(
    | venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | org_catalog/settings
    | org_catalog/docker
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.poetry]
name = "drf_organizations_catalog"
version = "0.1.0"
description = ""
authors = ["DRAGANmik <75855531+DRAGANmik@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"
drf-yasg2 = "^1.19.4"
black = "^21.8b0"
flake8 = "^3.9.2"
isort = "^5.9.3"
psycopg2-binary = "2.8.6"

[tool.poetry.dev-dependencies]
Django = "^3.2.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
